// src/models/WebSocketMessageType.ts
export enum WebSocketMessageType {
    NEW_MESSAGE = "NEW_MESSAGE",
    POPULARITY_UPDATE = "POPULARITY_UPDATE",
    BAN_USER = "BAN_USER",
    BAN_USER_CHAT_NOTIFICATION = "BAN_USER_CHAT_NOTIFICATION",
    WARN_USER = "WARN_USER",
    ROLE_UPDATE_NOTIFICATION = "ROLE_UPDATE_NOTIFICATION",
    DELETE_MESSAGE = "DELETE_MESSAGE",
    MESSAGE_REACTION_UPDATE = "MESSAGE_REACTION_UPDATE",
    MESSAGE_EDIT = "MESSAGE_EDIT",
    REPORT_NOTIFICATION = "REPORT_NOTIFICATION",
    CHATROOM_ARCHIVED = "CHATROOM_ARCHIVED",
    CHATROOM_UNARCHIVED = "CHATROOM_UNARCHIVED",
    MESSAGE_SENDING_AVAILABILITY_UPDATE = "MESSAGE_SENDING_AVAILABILITY_UPDATE",
    PRIVATE_NEW_MESSAGE = "PRIVATE_NEW_MESSAGE",
    PRIVATE_MESSAGE_EDIT = "PRIVATE_MESSAGE_EDIT",
    PRIVATE_MESSAGE_DELETE = "PRIVATE_MESSAGE_DELETE"
}